test: organise tests to mirror src file locations
