+++ test/integration/routes/kits-v1/organisation/lock-unlock.test.js
@@ -1,5 +1,5 @@
 import Hapi from '@hapi/hapi'
-import { organisation } from '../../../../src/routes/kits-v1/organisation.js'
+import { organisation } from '../../../../../src/routes/kits-v1/organisation.js'
 
 describe('Organisation Lock/Unlock', () => {
   let server
