+++ test/integration/routes/kits-v1/person/person.test.js
@@ -1,6 +1,6 @@
 import Hapi from '@hapi/hapi'
-import { person } from '../../../../src/routes/kits-v1/person.js'
-import { loadSchema } from '../../../../src/utils/validatePayload.js'
+import { person } from '../../../../../src/routes/kits-v1/person.js'
+import { loadSchema } from '../../../../../src/utils/validatePayload.js'
 
 describe('Fake Person', () => {
   let server, schema
